+++ src/modules/voxelformat/VolumeFormat.cpp
@@ -49,6 +49,7 @@
 #include "voxelformat/private/qubicle/QBTFormat.h"
 #include "voxelformat/private/qubicle/QEFFormat.h"
 #include "voxelformat/private/rooms/ThingFormat.h"
+#include "voxelformat/private/sandbox/VXBFormat.h"
 #include "voxelformat/private/sandbox/VXCFormat.h"
 #include "voxelformat/private/sandbox/VXMFormat.h"
 #include "voxelformat/private/sandbox/VXRFormat.h"
@@ -151,6 +152,14 @@ static const io::FormatDescription &sandboxVXM() {
 								   {"vxm"},
 								   {"VXMA", "VXMB", "VXMC", "VXM9", "VXM8", "VXM7", "VXM6", "VXM5", "VXM4", "VXM3"},
 								   VOX_FORMAT_FLAG_PALETTE_EMBEDDED | FORMAT_FLAG_SAVE};
+	return f;
+}
+
+static const io::FormatDescription &sandboxVXB() {
+	static io::FormatDescription f{"Sandbox VoxEdit Block",
+								   {"vxb"},
+								   {"VXB1"},
+								   VOX_FORMAT_FLAG_PALETTE_EMBEDDED};
 	return f;
 }
 
@@ -346,6 +355,7 @@ const io::FormatDescription *voxelLoad() {
 												 sandboxTilemap(),
 												 sandboxCollection(),
 												 sandboxVXM(),
+												 sandboxVXB(),
 												 blockbench(),
 												 sandboxVXR(),
 												 binvox(),
@@ -440,6 +450,8 @@ static core::SharedPtr<Format> getFormat(const io::FormatDescription &desc, uint
 			return core::make_shared<VXMFormat>();
 		} else if (ext == sandboxVXR().mainExtension()) {
 			return core::make_shared<VXRFormat>();
+		} else if (ext == sandboxVXB().mainExtension()) {
+			return core::make_shared<VXBFormat>();
 		} else if (ext == voxelMax().mainExtension()) {
 			return core::make_shared<VMaxFormat>();
 		} else if (ext == blockbench().mainExtension()) {