+++ src/modules/voxelformat/private/sandbox/VXBFormat.h
@@ -1,0 +1,27 @@
+/**
+ * @file
+ */
+
+#pragma once
+
+#include "voxelformat/Format.h"
+
+namespace voxelformat {
+
+/**
+ * @brief VXB files are block tiles
+ *
+ * @ingroup Formats
+ */
+class VXBFormat : public PaletteFormat {
+protected:
+	bool loadGroupsPalette(const core::String &filename, const io::ArchivePtr &archive,
+						   scenegraph::SceneGraph &sceneGraph, palette::Palette &palette,
+						   const LoadContext &ctx) override;
+	size_t loadPalette(const core::String &filename, const io::ArchivePtr &archive, palette::Palette &palette,
+					   const LoadContext &ctx) override;
+	bool saveGroups(const scenegraph::SceneGraph &sceneGraph, const core::String &filename,
+					const io::ArchivePtr &archive, const SaveContext &ctx) override;
+};
+
+} // namespace voxelformat