+++ src/modules/voxelformat/tests/VXBFormatTest.cpp
@@ -1,0 +1,16 @@
+/**
+ * @file
+ */
+
+#include "AbstractFormatTest.h"
+
+namespace voxelformat {
+
+class VXBFormatTest : public AbstractFormatTest {};
+
+TEST_F(VXBFormatTest, testLoad) {
+	scenegraph::SceneGraph sceneGraph;
+	testLoad(sceneGraph, "sandbox-block2.vxb", 1);
+}
+
+} // namespace voxelformat